+++ crates/README.md
@@ -27,6 +27,8 @@ The intention is that (much) longer term, we might be wanting to expose more of 
 * `qiskit-synthesis` is the crate for synthesis functionality.
 * `qiskit-circuit-library` is the crate for circuit library functions. It contains constructors or other
   circuit functionality that builds on the core circuit data model defined in `qiskit-circuit`.
+* `qiskit-util` is for small utility functions and data structures that are independent of all
+  Qiskit-specific objects, so it can be depended-on by both `quantum-info` and `circuit`.
 
 We use a structure with several crates in it for a couple of reasons:
