+++ crates/accelerate/Cargo.toml
@@ -25,6 +25,7 @@ itertools.workspace = true
 qiskit-circuit.workspace = true
 qiskit-transpiler.workspace = true
 qiskit-quantum-info.workspace = true
+qiskit-util.workspace = true
 nalgebra.workspace = true
 
 [dependencies.smallvec]
