+++ crates/accelerate/src/pauli_exp_val.rs
@@ -18,8 +18,8 @@ use pyo3::prelude::*;
 use pyo3::wrap_pyfunction;
 use rayon::prelude::*;
 
-use qiskit_circuit::getenv_use_multiple_threads;
-use qiskit_circuit::util::c64;
+use qiskit_util::complex::c64;
+use qiskit_util::getenv_use_multiple_threads;
 
 const PARALLEL_THRESHOLD: usize = 19;
