+++ crates/accelerate/src/results/marginalization.rs
@@ -19,7 +19,7 @@ use numpy::IntoPyArray;
 use numpy::{PyReadonlyArray1, PyReadonlyArray2, PyReadonlyArray3};
 use pyo3::IntoPyObjectExt;
 use pyo3::prelude::*;
-use qiskit_circuit::getenv_use_multiple_threads;
+use qiskit_util::getenv_use_multiple_threads;
 use rayon::prelude::*;
 
 fn marginalize<T: std::ops::AddAssign + Copy>(
