+++ crates/circuit/src/lib.rs
@@ -10,8 +10,6 @@
 // copyright notice, and modified files need to carry a notice indicating
 // that they have been altered from the originals.
 
-use std::env;
-
 pub mod annotation;
 pub mod bit;
 pub mod bit_locator;
@@ -38,7 +36,6 @@ pub mod parameter;
 pub mod parameter_table;
 pub mod register_data;
 pub mod slice;
-pub mod util;
 pub mod var_stretch_container;
 mod variable_mapper;
 pub mod vf2;
@@ -214,19 +211,6 @@ impl From<CapacityError> for PyErr {
     fn from(val: CapacityError) -> PyErr {
         PyRuntimeError::new_err(val.to_string())
     }
-}
-
-#[inline]
-pub fn getenv_use_multiple_threads() -> bool {
-    let parallel_context = env::var("QISKIT_IN_PARALLEL")
-        .unwrap_or_else(|_| "FALSE".to_string())
-        .to_uppercase()
-        == "TRUE";
-    let force_threads = env::var("QISKIT_FORCE_THREADS")
-        .unwrap_or_else(|_| "FALSE".to_string())
-        .to_uppercase()
-        == "TRUE";
-    !parallel_context || force_threads
 }
 
 pub fn circuit(m: &Bound<PyModule>) -> PyResult<()> {
