+++ crates/circuit_library/Cargo.toml
@@ -16,6 +16,7 @@ workspace = true
 qiskit-circuit.workspace = true
 qiskit-quantum-info.workspace = true
 qiskit-synthesis.workspace = true
+qiskit-util.workspace = true
 rayon.workspace = true
 num-complex.workspace = true
 nalgebra.workspace = true
