+++ crates/synthesis/Cargo.toml
@@ -15,6 +15,7 @@ workspace = true
 [dependencies]
 qiskit-circuit.workspace = true
 qiskit-quantum-info.workspace = true
+qiskit-util.workspace = true
 bytemuck.workspace = true
 num-complex.workspace = true
 nalgebra.workspace = true
