+++ crates/transpiler/src/passes/barrier_before_final_measurement.rs
@@ -110,19 +110,18 @@ pub fn run_barrier_before_final_measurements(
         nodes
     };
 
-    let final_ops: Vec<NodeIndex> = if dag.num_qubits() >= PARALLEL_THRESHOLD
-        && ::qiskit_circuit::getenv_use_multiple_threads()
-    {
-        dag.qubit_io_map()
-            .par_iter()
-            .flat_map(find_final_nodes)
-            .collect()
-    } else {
-        dag.qubit_io_map()
-            .iter()
-            .flat_map(find_final_nodes)
-            .collect()
-    };
+    let final_ops: Vec<NodeIndex> =
+        if dag.num_qubits() >= PARALLEL_THRESHOLD && ::qiskit_util::getenv_use_multiple_threads() {
+            dag.qubit_io_map()
+                .par_iter()
+                .flat_map(find_final_nodes)
+                .collect()
+        } else {
+            dag.qubit_io_map()
+                .iter()
+                .flat_map(find_final_nodes)
+                .collect()
+        };
 
     if final_ops.is_empty() {
         return Ok(());
